+++ src/frontends/sdl2/Main.cpp
@@ -14,8 +14,8 @@
 
 #include "AppConfig.h"
 #include "apple2/Video.h"
-#include "core/LinAppleCore.h"
 #include "apple2/peripherals/Peripheral_Audio.h"
+#include "core/LinAppleCore.h"
 #include "frontends/common/AppArgs.h"
 #include "frontends/common/AppController.h"
 #include "frontends/common/AudioDumper.h"
@@ -89,7 +89,7 @@ auto ds_init() -> bool {
       [](int slot) -> void { audio_mixer_unregister_source(slot); });
 
   linapple_set_audio_channel_callback(
-      [](const char* peripheral_id, int slot, const int16_t* const* channels,
+      [](const char* peripheral_id, int slot, const float* const* channels,
          size_t num_channels, size_t num_samples) -> void {
         audio_mixer_upload_channels(peripheral_id, slot, channels, num_channels,
                                     static_cast<uint32_t>(num_samples));
